refactor: add common input, output view class to seperate input, output logic from view logic
